build: bump Zappa to unreleased working version
The latest Zappa version cannot be installed with the latest setuptools. A PR with a fix (https://github.com/zappa/Zappa/pull/1029) has been merged in the Zappa project but there hasn't been a new release yet.

When Zappa releases a new version, this change can be reverted and replaced with a bump of the latest Zappa version.